separated the config stuff
